Split test_post_users into two files and pydoc
