+++ app/core/tests/test_post_users.py
@@ -1,17 +1,13 @@
 import pytest
 from django.urls import reverse
 from rest_framework import status
-from rest_framework.exceptions import ValidationError
-from rest_framework.test import APIClient
 from rest_framework.test import APIRequestFactory
 from rest_framework.test import force_authenticate
 
 from constants import global_admin
 from core.api.views import UserViewSet
 from core.field_permissions import FieldPermissions
-from core.permission_util import PermissionUtil
 from core.tests.utils.seed_constants import garry_name
-from core.tests.utils.seed_constants import wanda_project_lead
 from core.tests.utils.seed_user import SeedUser
 
 count_website_members = 4
@@ -37,53 +33,8 @@ class TestPostUser:
     def teardown_method(self):
         FieldPermissions.derive_cru_fields()
 
-    def test_admin_post_request_succeeds(self):  #
-        requester = SeedUser.get_user(garry_name)
-        client = APIClient()
-        client.force_authenticate(user=requester)
-
-        url = reverse("user-list")
-        data = {
-            "username": "createuser",
-            "last_name": "created",
-            "gmail": "create@example.com",
-            "password": "password",
-        }
-        response = client.post(url, data, format="json")
-        assert response.status_code == status.HTTP_201_CREATED
-
-    def test_admin_post_with_created_at_fails(self):
-        requester = SeedUser.get_user(garry_name)
-        client = APIClient()
-        client.force_authenticate(user=requester)
-
-        url = reverse("user-list")
-        data = {
-            "username": "createuser",
-            "last_name": "created",
-            "gmail": "create@example.com",
-            "password": "password",
-            "time_zone": "America/Los_Angeles",
-            "created_at": "2022-01-01T00:00:00Z",
-        }
-        response = client.post(url, data, format="json")
-        assert response.status_code == status.HTTP_400_BAD_REQUEST
-
-    def test_validate_fields_postable_raises_exception_for_created_at(self):
-        with pytest.raises(ValidationError):
-            PermissionUtil.validate_fields_postable(
-                SeedUser.get_user(garry_name),
-                ["created_at"],
-            )
-
-    def test_validate_fields_postable_raises_exception_for_project_lead(self):
-        with pytest.raises(PermissionError):
-            PermissionUtil.validate_fields_postable(
-                SeedUser.get_user(wanda_project_lead), ["username", "password"]
-            )
-
     def test_allowable_post_fields_configurable(self):
-        """Test that the fields that can be updated are configurable.
+        """Test POST request returns success when the request fields match configured fields.
 
         This test mocks a PATCH request to skip submitting the request to the server and instead
         calls the view directly with the request.  This is done so that variables used by the
@@ -116,7 +67,10 @@ class TestPostUser:
         assert response.status_code == status.HTTP_201_CREATED
 
     def test_not_allowable_post_fields_configurable(self):
-        """Test that the fields that are not configured to be updated cannot be updated.
+        """Test post request returns 400 response when request fields do not match configured fields.
+
+        Fields are configured to not include last_name.  The test will attempt to create a user
+        with last_name in the request data.  The test should fail with a 400 status code.
 
         See documentation for test_allowable_patch_fields_configurable for more information.
         """
@@ -131,7 +85,6 @@ class TestPostUser:
         ]
 
         requester = SeedUser.get_user(garry_name)  # project lead for website
-
         post_data = {
             "username": "foo",
             "last_name": "Smith",
